refactor: split 3 oversized files + complete PMAT-077 diagnostics (Refs PMAT-077)
- explain.rs 972→434 lines (extracted explain_render.rs + explain_tests.rs)
- lint/config.rs 561→282 lines (extracted config_tests.rs)
- schema/validator_tests.rs 543→278 lines (extracted validator_tests_extra.rs)

All files now under 500 lines. 14/14 lint PASS, 13/13 builds OK.
