+++ .pmat-work/PMAT-077/rescue/rescue-54f3bc22-f291-4685-b988-0cf8966aade8.json
@@ -1,0 +1,21 @@
+{
+  "rescue_id": "54f3bc22-f291-4685-b988-0cf8966aade8",
+  "work_item_id": "PMAT-077",
+  "timestamp": "2026-03-25T13:34:56.048015685Z",
+  "violated_clause": "All changes pushed",
+  "strategy": {
+    "ManualIntervention": {
+      "guidance": "Push all local commits: `git push origin HEAD`"
+    }
+  },
+  "diagnosis": {
+    "summary": "Manual intervention required: Push all local commits: `git push origin HEAD`",
+    "suggested_actions": [
+      "Push all local commits: `git push origin HEAD`",
+      "Re-run: pmat work complete <id>"
+    ],
+    "generated_artifacts": []
+  },
+  "outcome": "ManualActionRequired",
+  "retry_allowed": true
+}
