+++ .pmat-work/PMAT-077/rescue/rescue-bb88b4fa-39de-455f-9f89-a3ec2461ca42.json
@@ -1,0 +1,21 @@
+{
+  "rescue_id": "bb88b4fa-39de-455f-9f89-a3ec2461ca42",
+  "work_item_id": "PMAT-077",
+  "timestamp": "2026-03-25T13:34:56.047915143Z",
+  "violated_clause": "No vulnerable dependencies added",
+  "strategy": {
+    "ManualIntervention": {
+      "guidance": "Run `cargo audit` to identify vulnerable dependencies, then update or replace them."
+    }
+  },
+  "diagnosis": {
+    "summary": "Manual intervention required: Run `cargo audit` to identify vulnerable dependencies, then update or replace them.",
+    "suggested_actions": [
+      "Run `cargo audit` to identify vulnerable dependencies, then update or replace them.",
+      "Re-run: pmat work complete <id>"
+    ],
+    "generated_artifacts": []
+  },
+  "outcome": "ManualActionRequired",
+  "retry_allowed": true
+}
