+++ .pmat-work/PMAT-077/rescue/rescue-eea7dafa-d593-46e8-9509-7a71022cedb0.json
@@ -1,0 +1,21 @@
+{
+  "rescue_id": "eea7dafa-d593-46e8-9509-7a71022cedb0",
+  "work_item_id": "PMAT-077",
+  "timestamp": "2026-03-25T13:34:56.047986491Z",
+  "violated_clause": "No file exceeds 500 lines",
+  "strategy": {
+    "ManualIntervention": {
+      "guidance": "Split large files using extract-module refactoring. Target files exceeding the line limit."
+    }
+  },
+  "diagnosis": {
+    "summary": "Manual intervention required: Split large files using extract-module refactoring. Target files exceeding the line limit.",
+    "suggested_actions": [
+      "Split large files using extract-module refactoring. Target files exceeding the line limit.",
+      "Re-run: pmat work complete <id>"
+    ],
+    "generated_artifacts": []
+  },
+  "outcome": "ManualActionRequired",
+  "retry_allowed": true
+}
