+++ .pmat-work/ledger.jsonl
@@ -193,3 +193,7 @@
 {"receipt_id":"019cc983-f235-7913-9b47-a572fbeba030","work_item_id":"PMAT-073","timestamp":"2026-03-07T18:16:33.589781045+00:00","git_sha":"97ed193a8d54defa734312d0528b2cba3c1971e9","trigger":"WorkComplete","passed":17,"failed":4,"overridden":4,"allows_completion":true,"content_hash":"33f50ad360cd062ae8ab75938132e00ba0e0a0719742e9f1db8171a581fdcaf5"}
 {"receipt_id":"019cc984-4537-7b01-a633-5cf7cc036c2d","work_item_id":"PMAT-073","timestamp":"2026-03-07T18:16:54.839310573+00:00","git_sha":"f3b60c1ba3d9751f47e5f20c057d8fbe533bac87","trigger":"WorkComplete","passed":17,"failed":4,"overridden":4,"allows_completion":true,"content_hash":"4f3d9caaaf3b29f379b493339553c62b53e81fffd79ecee7b66740bc1463038a"}
 {"receipt_id":"019cc98a-1cb1-7872-83ad-28b709406d4e","work_item_id":"PMAT-074","timestamp":"2026-03-07T18:23:17.681664687+00:00","git_sha":"036e349fa3a0b4d576a7f6ddb6f814847ec34dc8","trigger":"WorkComplete","passed":18,"failed":3,"overridden":3,"allows_completion":true,"content_hash":"0dbc7c4c5e724d2f943ed4b0dc38d1dc23ddfb3c728a804938103442331ccb0b"}
+{"receipt_id":"019d2533-5d6b-7ad2-8979-df1635c6d60b","work_item_id":"PMAT-077","timestamp":"2026-03-25T13:33:36.491659211+00:00","git_sha":"521793dd97b5fb0786908d52bb6a226823531de6","trigger":"WorkComplete","passed":17,"failed":4,"overridden":3,"allows_completion":false,"content_hash":"6b099654a46cb5576c8895cf165007a8103f569778c47bd63e5f87b87c298baf"}
+{"receipt_id":"019d2533-c8ea-7350-829d-23abd0ad6403","work_item_id":"PMAT-077","timestamp":"2026-03-25T13:34:04.010369724+00:00","git_sha":"521793dd97b5fb0786908d52bb6a226823531de6","trigger":"WorkComplete","passed":16,"failed":5,"overridden":4,"allows_completion":false,"content_hash":"d4dd020fad1e82bd0556692399845485584018d1a15c9eba202a0135c7432aa7"}
+{"receipt_id":"019d2534-942f-7680-a97b-66d1726c59d8","work_item_id":"PMAT-077","timestamp":"2026-03-25T13:34:56.047690383+00:00","git_sha":"521793dd97b5fb0786908d52bb6a226823531de6","trigger":"WorkComplete","passed":16,"failed":5,"overridden":0,"allows_completion":false,"content_hash":"37c6762c1af3ff1456cbc5dbefee5145f4b2517e7947e2a6b7c9862b2110e5de"}
+{"receipt_id":"019d2534-fd38-7b80-a470-c87d906c424c","work_item_id":"PMAT-077","timestamp":"2026-03-25T13:35:22.936644306+00:00","git_sha":"521793dd97b5fb0786908d52bb6a226823531de6","trigger":"WorkComplete","passed":16,"failed":5,"overridden":0,"allows_completion":false,"content_hash":"891bbd535e73d24e0d3254806f216361727a72cde3036cfe5a70daa054c623a4"}
