+++ .pmat/project.toml
@@ -1,4 +1,4 @@
 [pmat]
 version = "3.8.0"
-last_compliance_check = "2026-03-25T13:29:15.724229880Z"
+last_compliance_check = "2026-03-25T13:42:26.392712468Z"
 auto_update = false
